GCP/CT/requirements.txt
google-cloud-storage==2.7.0
scikit-learn==1.2.2
pandas==1.4.4
numpy==1.22.4
Pillow==8.4.0 
google-auth-oauthlib==0.4.6